+++ internal/config/config.go
@@ -10,14 +10,7 @@ import (
 	"github.com/nao1215/gup/internal/cmdinfo"
 	"github.com/nao1215/gup/internal/file"
 	"github.com/nao1215/gup/internal/goutil"
-	"github.com/nao1215/gup/internal/print"
 )
-
-func init() {
-	if err := os.MkdirAll(DirPath(), 0775); err != nil {
-		print.Err(fmt.Errorf("%s: %w", "can not make config directory", err))
-	}
-}
 
 // FilePath return configuration-file path.
 func FilePath() string {
@@ -26,7 +19,15 @@ func FilePath() string {
 
 // DirPath return directory path that store configuration-file.
 func DirPath() string {
-	return filepath.Join(os.Getenv("HOME"), ".config", cmdinfo.Name())
+	home, err := os.UserHomeDir()
+	if err != nil {
+		// If $HOME is empty, .config directory can be created in the
+		// current directory. The .config directory path is displayed
+		// when reporting the completion of the export subcommand.
+		// So, user notices that the output destination is strange.
+		return filepath.Join(os.Getenv("HOME"), ".config", cmdinfo.Name())
+	}
+	return filepath.Join(home, ".config", cmdinfo.Name())
 }
 
 // ReadConfFile return contents of configuration-file (package information)